+++ src/components/App/App.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import GuestLoves from '../GuestLoves/GuestLoves';
+import './App.css';
+
+function App() {
+	return (
+		<>
+			<GuestLoves />
+		</>
+	);
+}
+
+export default App;
