+++ src/components/GuestLoves/guestLoves.css
@@ -1,0 +1,19 @@
+.guest-loves {
+    padding: 120px 0;
+}
+
+.guest-loves__slider-wrapper {
+    position: relative;
+}
+
+.guest-loves__swiper-container {
+    position: relative;
+    overflow: hidden;
+    width: 100%;
+}
+
+.guest-loves__swiper-wrapper {
+    display: flex;
+    width: 100%;
+    height: 100%;
+}
