--- src/components/guestLoves.css
@@ -1,58 +1,0 @@
-.sprites {
-    display: none;
-}
-
-.container {
-    width: 100%;
-    max-width: 1232px;
-    margin: 0 auto;
-}
-
-
-h2 {
-    font-weight: 500;
-    font-size: 40px;
-    text-align: center;
-    padding-top: 120px;
-}
-
-.guests {
-    background-color: #eaf0f9;
-    position: relative;
-}
-
-.guests-loves__title {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    padding-bottom: 106px;
-}
-
-.guests-loves__title h2 {
-    color: #383838;
-}
-
-
-
-.guest-loves {
-    padding: 120px 0;
-}
-
-.guest-loves__slider-wrapper {
-    position: relative;
-}
-
-
-
-.guest-loves__swiper-container {
-    position: relative;
-    overflow: hidden;
-    width: 100%;
-}
-
-.guest-loves__swiper-wrapper {
-    display: flex;
-    width: 100%;
-    height: 100%;
-}
-
